requirements.txt
SQLAlchemy~=2.0.40
pydantic~=2.11.4
python-dotenv~=1.1.0
fastapi~=0.115.12
starlette~=0.46.2
uvicorn~=0.34.2
asyncpg~=0.30.0